Separate FlowTable datatypes from implementation
